Avoid logging non-exceptional exception, and fix error in log format string
If loading data types out of order it is expected to get an exception in load_data_typ_definitions. We shouldn't report this to the log file until we give up on loading types.

This also fixes an error in the log format string, which leads to an exception when logging.